Misc fixes

+++ sources/addons2.php
@@ -707,12 +707,12 @@ function uninstall_addon($addon)
 
                 $matches = array();
                 if (preg_match('#([^/]*)/?pages/modules(_custom)?/(.*)\.php#', $filename, $matches) != 0) {
-                    if ($matches[2] != '_custom' || !is_file(get_file_base() . '/' . $filename)) {
+                    if ($matches[2] != '_custom' || ($matches[2] == '_custom' && !is_file(get_file_base() . '/' . str_replace('_custom/', '/', $filename)))) {
                         uninstall_module($matches[1], $matches[3]);
                     }
                 }
                 if (preg_match('#sources(_custom)?/blocks/(.*)\.php#', $filename, $matches) != 0) {
-                    if ($matches[1] != '_custom' || !is_file(get_file_base() . '/' . $filename)) {
+                    if ($matches[1] != '_custom' || ($matches[1] == '_custom' && !is_file(get_file_base() . '/' . str_replace('_custom/', '/', $filename)))) {
                         uninstall_block($matches[2]);
                     }
                 }
